Extract Activity helpers from DashboardHelper to Shared::ActivitiesHelper
