+++ app/controllers/locomotive/dashboard_controller.rb
@@ -1,5 +1,6 @@
 module Locomotive
   class DashboardController < BaseController
+    include Locomotive::Shared::ActivitiesHelper
 
     account_required & within_site
